add: qwen 3.5
