+++ src/axolotl/monkeypatch/multipack.py
@@ -22,6 +22,8 @@ SUPPORTED_MULTIPACK_MODEL_TYPES = [
     "qwen3",
     "qwen3_moe",
     "qwen3_next",
+    "qwen3_5",
+    "qwen3_5_moe",
     "falcon",
     "phi",
     "phi3",
